modify reducers
