+++ project/src/store/comments-process/selectors.ts
@@ -1,0 +1,8 @@
+import {NameSpace} from '../../const';
+import {State} from '../../types/state';
+
+export const getReviewBlockData = (state: State) => ({
+  comments: state[NameSpace.Comments],
+  authorizationStatus: state[NameSpace.User].authorizationStatus,
+});
+
