+++ project/src/store/offers-nearby-process/offers-nearby-process.ts
@@ -1,5 +1,5 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
-import {setRoomData} from './room-reducer';
+import {setRoomData} from '../room-process/room-process';
 import {Offer, RoomDataType} from '../../types/offers';
 import {NameSpace} from '../../const';
