+++ project/src/store/store.ts
@@ -1,6 +1,6 @@
 import {configureStore} from '@reduxjs/toolkit';
 import {createAPI} from '../services/api';
-import reducer from './reducers/reducers';
+import reducer from './root-reducer';
 import {redirect} from './middlewares/redirect';
 
 const api = createAPI();
